requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.4
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
idna==3.11
packaging==26.1
pillow==12.2.0
PyJWT==2.12.1
qrcode==8.2
redis==7.4.0
requests==2.33.1
sqlparse==0.5.5
urllib3==2.6.3
